Outlining Architecture in an attempt to follow an Onion Layer approach.
Cultivate_DataFramework will be the Respository layer, housing the different interfaces for connecting to and retreiving data from different DBMSes.

Cultivate_MealPlanning, Cultivate_Recipes, Cultivate_Inventory, and Cultivate_Utilities will be the Buisness Logic layer, handling the behaviour of the domain entities. MealPlanning, Recipes and Inventory are all separate projects in an attempt to divide components up and keep them as strictly segragated as possible.

Cultivate_WindowsDekstop is currently the only UI component in the Client layer. This project will attempt to have very little if any logic in it, other than displaying objects to the user and sending changes back up the dependency tree.